refacor: WalletVector and UnhardenedWalletVector in a new file
